+++ src/test/kotlin/org/jetbrains/intellij/platform/gradleplugin/tasks/BuildPluginTaskSpec.kt
@@ -95,7 +95,7 @@ class BuildPluginTaskSpec : IntelliJPluginSpecBase() {
             """
             <idea-plugin>
               <version>0.42.123</version>
-              <idea-version since-build="212.5712" until-build="212.*" />
+              <idea-version since-build="221.6008" until-build="221.*" />
               <name>MyPluginName</name>
               <vendor>JetBrains</vendor>
               <depends config-file="other.xml" />