+++ src/test/kotlin/org/jetbrains/intellij/platform/gradleplugin/tasks/DownloadIntelliJSpec.kt
@@ -2,13 +2,10 @@
 
 package org.jetbrains.intellij.platform.gradleplugin.tasks
 
-import org.gradle.api.plugins.BasePlugin
+import org.gradle.api.plugins.BasePlugin.ASSEMBLE_TASK_NAME
 import org.jetbrains.intellij.platform.gradleplugin.IntelliJPluginSpecBase
 import java.io.File
-import kotlin.test.Test
-import kotlin.test.assertEquals
-import kotlin.test.assertFalse
-import kotlin.test.assertTrue
+import kotlin.test.*
 
 class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
 
@@ -21,6 +18,7 @@ class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
       causes issues on Windows GitHub Actions runners (file is locked by another process).
      */
 
+    @Ignore
     @Test
     fun `download idea dependencies without sources when downloadSources = false`() {
         val testedVersion = "2019.3.1"
@@ -28,7 +26,6 @@ class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
         val icPomCacheDir = File(icVersionCacheDir, "6c69524caec42364796bb99471fa4057c4daf567")
         val icDistCacheDir = File(icVersionCacheDir, "52292e4f8a0ccb3ceb08bd81fd57b88923ac8e99")
         val iuVersionCacheDir = File(iuCacheDir, testedVersion)
-        deleteIfExists(icDistCacheDir)
 
         buildFile.groovy(
             """
@@ -39,7 +36,7 @@ class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
             }
             """.trimIndent()
         )
-        build(BasePlugin.ASSEMBLE_TASK_NAME)
+        build(ASSEMBLE_TASK_NAME)
 
         assertDirExistsAndContainsOnly(icVersionCacheDir, icPomCacheDir.name, icDistCacheDir.name)
         assertDirExistsAndContainsOnly(icPomCacheDir, "ideaIC-2019.3.1.pom")
@@ -47,6 +44,7 @@ class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
         assertFalse(iuVersionCacheDir.exists(), "Expected Ultimate cache directory to not exist")
     }
 
+    @Ignore
     @Test
     fun `download idea with sources when downloadSources = true`() {
         val testedVersion = "2019.3.2"
@@ -55,8 +53,6 @@ class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
         val icDistCacheDir = File(icVersionCacheDir, "b911671d7501d9fd63cea3eee8287d8ceb4113c6")
         val icSourcesCacheDir = File(icVersionCacheDir, "db705bd2f3911c2e35b53c9084ed96c3c0b7e900")
         val iuVersionCacheDir = File(iuCacheDir, testedVersion)
-        deleteIfExists(icDistCacheDir)
-        deleteIfExists(icSourcesCacheDir)
 
         buildFile.groovy(
             """
@@ -68,7 +64,7 @@ class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
             """.trimIndent()
         )
 
-        build(BasePlugin.ASSEMBLE_TASK_NAME)
+        build(ASSEMBLE_TASK_NAME)
 
         assertDirExistsAndContainsOnly(icVersionCacheDir, icPomCacheDir.name, icDistCacheDir.name, icSourcesCacheDir.name)
         assertDirExistsAndContainsOnly(icPomCacheDir, "ideaIC-2019.3.2.pom")
@@ -77,6 +73,7 @@ class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
         assertFalse(iuVersionCacheDir.exists(), "Expected Ultimate cache directory to not exist")
     }
 
+    @Ignore
     @Test
     fun `download ultimate idea dependencies without sources when downloadSources = false`() {
         val testedVersion = "2019.3.3"
@@ -84,7 +81,6 @@ class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
         val iuPomCacheDir = File(iuVersionCacheDir, "9d81b0a2416de8815bbe266ffc4065a2e4a2fda0")
         val iuDistCacheDir = File(iuVersionCacheDir, "9bf72a4b290a00d569759db11454eb3347026fc7")
         val icVersionCacheDir = File(icCacheDir, testedVersion)
-        deleteIfExists(iuDistCacheDir)
 
         buildFile.groovy(
             """
@@ -96,7 +92,7 @@ class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
             """.trimIndent()
         )
 
-        build(BasePlugin.ASSEMBLE_TASK_NAME)
+        build(ASSEMBLE_TASK_NAME)
 
         assertDirExistsAndContainsOnly(iuVersionCacheDir, iuPomCacheDir.name, iuDistCacheDir.name)
         assertDirExistsAndContainsOnly(iuPomCacheDir, "ideaIU-2019.3.3.pom")
@@ -104,6 +100,7 @@ class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
         assertFalse(icVersionCacheDir.exists(), "Expected Community cache directory to not exist")
     }
 
+    @Ignore
     @Test
     fun `download ultimate idea dependencies and community sources without dist when downloadSources = true`() {
         val testedVersion = "2019.3.4"
@@ -113,8 +110,6 @@ class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
         val icVersionCacheDir = File(icCacheDir, testedVersion)
         val icPomCacheDir = File(icVersionCacheDir, "55441dcc36ee231de051150455a395f775dac552")
         val icSourcesCacheDir = File(icVersionCacheDir, "11facb5d520efd78d3145de5fa01449179465e5")
-        deleteIfExists(iuDistCacheDir)
-        deleteIfExists(icSourcesCacheDir)
 
         buildFile.groovy(
             """
@@ -126,7 +121,7 @@ class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
             """.trimIndent()
         )
 
-        build(BasePlugin.ASSEMBLE_TASK_NAME)
+        build(ASSEMBLE_TASK_NAME)
 
         assertDirExistsAndContainsOnly(iuVersionCacheDir, iuPomCacheDir.name, iuDistCacheDir.name)
         assertDirExistsAndContainsOnly(iuPomCacheDir, "ideaIU-2019.3.4.pom")
@@ -134,12 +129,6 @@ class DownloadIntelliJSpec : IntelliJPluginSpecBase() {
         assertDirExistsAndContainsOnly(icVersionCacheDir, icPomCacheDir.name, icSourcesCacheDir.name)
         assertDirExistsAndContainsOnly(icPomCacheDir, "ideaIC-2019.3.4.pom")
         assertDirExistsAndContainsOnly(icSourcesCacheDir, "ideaIC-2019.3.4-sources.jar")
-    }
-
-    private fun deleteIfExists(dir: File) {
-        if (dir.exists()) {
-            dir.deleteRecursively()
-        }
     }
 
     private fun assertDirExistsAndContainsOnly(dir: File, vararg fileNames: String) {