+++ src/test/kotlin/org/jetbrains/intellij/platform/gradleplugin/tasks/RunPluginVerifierTaskSpec.kt
@@ -356,7 +356,7 @@ class RunPluginVerifierTaskSpec : IntelliJPluginSpecBase() {
             class App {
             
                 public static void main(@NotNull String[] strings) {
-                    StringUtil.firstLetterToUpperCase("foo");
+                    StringUtil.escapeXml("<foo>");
                 }
             }
             """.trimIndent()