+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 - Configure all tasks that extend task classes instead of just those created by the plugin
 - Make JbrResolver prefer Gradle javaToolchains by `JetBrains s.r.o`, if available. Only otherwise start fetching and running a new one.
+- Support for Kotlin Coroutines debugging
+- Detect and warn if project adds an explicit dependency on Kotlin Coroutines library
 - Added `org.jetbrains.intellij.buildFeature.useCacheRedirector` build feature
 - Added `IntelliJPlatformCollectorTransformer` and `IntelliJPlatformExtractTransformer` for handling IntelliJ Platform dependency archives
 - Added `jetbrainsAnnotations` dependency helper for adding JetBrains Annotations dependency